change for double to work properly

+++ Main.java
@@ -7,15 +7,15 @@ class Main{
       Scanner input = new Scanner(System.in);
   System.out.println("Enter operation (and, or, xor, not, nand, nor, xnor):");
       String operation=input.next();
-      int val1=0, val2=0;
+      double val1=0, val2=0;
       if (operation.toLowerCase().equals("not")) {
         System.out.println("Enter a single value:");
-         val1 = input.nextInt();
+         val1 = input.nextDouble();
       } else{
         System.out.println("Enter the first value:");
-         val1=input.nextInt();
+         val1=input.nextDouble();
         System.out.println("Enter the second value:");
-         val2=input.nextInt();
+         val2=input.nextDouble();
       }
   input.close();
